Fixes as discussed in pull request: separate file and improve message
